upgrade to sokol 2023.06 in header-only form
No longer try to use sokol as a library using the libsokol-gfx lhelper
recipe but instead use sokol directly as a header only library.

The application sokol-shdc is not obtained from lhelper using the recipe
sokol-tools-bin that fetch directly the binary provided by the sokol
author.

We switch also to a new version of sokol so a few small changes have
been made so that it compiles correctly with the new version.